fix: rename library artifact and package structure
