+++ src/main/java/com/techconative/posf/modal/OrderingCriteria.java
@@ -1,4 +1,4 @@
-package com.techconative.inmemory.pagination.modal;
+package com.techconative.posf.modal;
 
 /** Order Criteria for sorting of data */
 public enum OrderingCriteria {
