--- src/test/java/com/techconative/inmemory/pagination/service/PaginationServiceTest.java
@@ -1,3 +1,0 @@
-package com.techconative.inmemory.pagination.service;
-
-public class PaginationServiceTest {}
