+++ src/test/java/com/techconative/posf/RunCucumberTest.java
@@ -1,4 +1,4 @@
-package com.techconative.inmemory.pagination;
+package com.techconative.posf;
 
 import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
 import static io.cucumber.junit.platform.engine.Constants.PLUGIN_PROPERTY_NAME;
@@ -10,7 +10,7 @@ import org.junit.platform.suite.api.Suite;
 
 @Suite
 @IncludeEngines("cucumber")
-@SelectClasspathResource("cucumber/pagination")
+@SelectClasspathResource("cucumber/posf")
 @ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "pretty")
-@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "com.techconative.inmemory.pagination")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "com.techconative.posf")
 public class RunCucumberTest {}
