+++ src/test/java/com/techconative/posf/StudentService.java
@@ -1,16 +1,17 @@
-package com.techconative.inmemory.pagination;
+package com.techconative.posf;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.techconative.inmemory.pagination.modal.Student;
-import com.techconative.inmemory.pagination.service.PaginationService;
+import com.techconative.posf.modal.Student;
+import com.techconative.posf.service.POSFService;
 import java.io.IOException;
 import java.util.List;
+
 import lombok.extern.slf4j.Slf4j;
 
 /** StudentService implements library for processing src/test/java/resources/Student.json */
 @Slf4j
-public class StudentService extends PaginationService<Student> {
+public class StudentService extends POSFService<Student> {
 
     private static ObjectMapper mapper = new ObjectMapper();
